chore: update all go builds to 1.25.1
This brings the Dockerfile and `go.mod` up-to-date
with the `stable` designation we use in our GHA
workflows.

This change also includes some Docker hygiene updates
with the includion of `.dockerignore` and a `README.md`
with clear instructions on building and running the image.

Finally, a basic CI workflow is added for the Dockerfile
to ensure it receives better maintenance/attention moving
forward.

Signed-off-by: Travis Truman <trumant@gmail.com>